+++ app/src/components/Map/PolygonReport.jsx
@@ -5,34 +5,7 @@ import classnames from 'classnames';
 import PolygonReportStyles from 'styles/components/map/c-polygon-report.scss';
 import buttonCloseStyles from 'styles/components/c-button-close.scss';
 import CloseIcon from 'babel!svg-react!assets/icons/close.svg?name=Icon';
-
-class PolygonReportInfoWindow extends google.maps.OverlayView {
-  constructor() {
-    super();
-    this.latLng = { lat: 0, lng: 0 };
-    this.div = document.createElement('div');
-    this.div.style.borderStyle = 'none';
-    this.div.style.borderWidth = '0px';
-    this.div.style.position = 'absolute';
-  }
-
-  onAdd() {
-    const panes = this.getPanes();
-    panes.floatPane.appendChild(this.div);
-  }
-
-  draw() {
-    const overlayProjection = this.getProjection();
-    const sw = overlayProjection.fromLatLngToDivPixel(new google.maps.LatLng(this.latLng));
-    this.div.style.left = `${sw.x}px`;
-    this.div.style.top = `${sw.y}px`;
-  }
-
-  setLatLng(latLng) {
-    this.latLng = { lat: latLng[0], lng: latLng[1] };
-    this.draw();
-  }
-}
+import CustomInfoWindow from 'util/CustomInfoWindow';
 
 export default class PolygonReport extends Component {
   constructor() {
@@ -42,7 +15,7 @@ export default class PolygonReport extends Component {
 
   componentWillReceiveProps(nextProps) {
     if (!this.map && nextProps.map) {
-      this.infoWindow = new PolygonReportInfoWindow();
+      this.infoWindow = new CustomInfoWindow();
       this.infoWindow.setMap(nextProps.map);
       this.infoWindow.div.addEventListener('click', this.onInfoWindowClickBound);
       this.map = nextProps.map;
